Isolate IIIF URL adapters
