+++ tests/live_dezoomers.rs
@@ -1,4 +1,4 @@
-fn run_live_dezoomer(name: &str, url: &str) {
+fn run_live_dezoomer(name: &str, url: &str, extra_args: &[&str]) {
     if std::env::var_os("DEZOOMIFY_LIVE_TESTS").is_none() {
         return;
     }
@@ -6,7 +6,7 @@ fn run_live_dezoomer(name: &str, url: &str) {
     let temp_dir = tempfile::tempdir().unwrap();
     let output = temp_dir.path().join(format!("{name}.png"));
     let mut command = std::process::Command::new(env!("CARGO_BIN_EXE_dezoomify-rs"));
-    command.args([url, output.to_str().unwrap(), "--max-width", "1000"]);
+    command.args([url, output.to_str().unwrap(), "--max-width", "1200"]);
     let result = command
         .args([
             "--retries",
@@ -20,6 +20,7 @@ fn run_live_dezoomer(name: &str, url: &str) {
             "--timeout",
             "30s",
         ])
+        .args(extra_args)
         .output()
         .unwrap();
     assert!(
@@ -32,10 +33,10 @@ fn run_live_dezoomer(name: &str, url: &str) {
 }
 
 macro_rules! live_dezoomer {
-    ($name:ident, $url:literal) => {
+    ($name:ident, $url:literal $(, $arg:literal)*) => {
         #[test]
         fn $name() {
-            run_live_dezoomer(stringify!($name), $url);
+            run_live_dezoomer(stringify!($name), $url, &[$($arg),*]);
         }
     };
 }
@@ -57,9 +58,15 @@ live_dezoomer!(
     iiif_csntm,
     "https://collections.csntm.org/image-service/iiif/MNTGRCGA01/default/M_NT_GRC_GA01_20250609_203r/M_NT_GRC_GA01_20250609_203r/info.json"
 );
+live_dezoomer!(iiif_onb_viewer, "https://viewer.onb.ac.at/10048A37/");
 live_dezoomer!(
-    iiif_onb_manifest,
-    "https://api.onb.ac.at/iiif/presentation/v3/manifest/10048A37"
+    iiif_oklahoma_contentdm,
+    "https://dc.library.okstate.edu/digital/collection/OKMaps/id/6483/rec/6",
+    "--accept-invalid-certs"
+);
+live_dezoomer!(
+    iiif_liechtenstein_collections,
+    "https://www.liechtensteincollections.at/en/collections-online/forest-landscape"
 );
 live_dezoomer!(
     iiif_nls_auchinleck,
